+++ x/wasm/keeper/msg_server_integration_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
 	"github.com/Finschia/finschia-sdk/testutil/testdata"
@@ -436,5 +437,110 @@ func TestExecuteContract(t *testing.T) {
 
 			require.NoError(t, err)
 		})
+	}
+}
+
+func TestClearAdmin(t *testing.T) {
+	wasmApp := app.Setup(false)
+	ctx := wasmApp.BaseApp.NewContext(false, tmproto.Header{Time: time.Now()})
+
+	var (
+		myAddress       sdk.AccAddress = make([]byte, types.ContractAddrLen)
+		_, _, otherAddr                = testdata.KeyTestPubAddr()
+	)
+
+	// setup
+	_, _, sender := testdata.KeyTestPubAddr()
+	storeMsg := types.MsgStoreCodeFixture(func(m *types.MsgStoreCode) {
+		m.Sender = sender.String()
+		m.WASMByteCode = wasmContract
+	})
+	rsp, err := wasmApp.MsgServiceRouter().Handler(storeMsg)(ctx, storeMsg)
+	require.NoError(t, err)
+	var storeCodeResult types.MsgStoreCodeResponse
+	require.NoError(t, wasmApp.AppCodec().Unmarshal(rsp.Data, &storeCodeResult))
+	codeID := storeCodeResult.CodeID
+
+	initMsg := types.MsgInstantiateContractFixture(func(m *types.MsgInstantiateContract) {
+		m.Sender = myAddress.String()
+		m.Admin = myAddress.String()
+		m.CodeID = codeID
+		m.Label = "test"
+		m.Msg = []byte(`{}`)
+		m.Funds = sdk.Coins{}
+	})
+	rsp, err = wasmApp.MsgServiceRouter().Handler(initMsg)(ctx, initMsg)
+	require.NoError(t, err)
+
+	var instantiateContractResult types.MsgInstantiateContractResponse
+	require.NoError(t, wasmApp.AppCodec().Unmarshal(rsp.Data, &instantiateContractResult))
+	contractAddress := instantiateContractResult.Address
+
+	specs := map[string]struct {
+		addr      string
+		expErr    bool
+		expEvents []abci.Event
+	}{
+		"admin can clear admin": {
+			addr:   myAddress.String(),
+			expErr: false,
+			expEvents: []abci.Event{
+				createMsgEvent(myAddress),
+				{
+					Type: "update_contract_admin",
+					Attributes: []abci.EventAttribute{
+						{
+							Key:   []byte("_contract_address"),
+							Value: []byte(contractAddress),
+						},
+						{
+							Key:   []byte("new_admin_address"),
+							Value: []byte{},
+						},
+					},
+				},
+			},
+		},
+		"other address cannot clear admin": {
+			addr:   otherAddr.String(),
+			expErr: true,
+		},
+	}
+	for name, spec := range specs {
+		t.Run(name, func(t *testing.T) {
+			xCtx, _ := ctx.CacheContext()
+			// when
+			msgClearAdmin := &types.MsgClearAdmin{
+				Sender:   spec.addr,
+				Contract: contractAddress,
+			}
+			rsp, err = wasmApp.MsgServiceRouter().Handler(msgClearAdmin)(xCtx, msgClearAdmin)
+
+			// then
+			if spec.expErr {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, spec.expEvents, rsp.Events)
+		})
+	}
+}
+
+// This function is utilized to generate the msg event for event checking in integration tests
+// It will be deleted in release/v0.1.x
+func createMsgEvent(sender sdk.AccAddress) abci.Event {
+	return abci.Event{
+		Type: "message",
+		Attributes: []abci.EventAttribute{
+			{
+				Key:   []byte("module"),
+				Value: []byte("wasm"),
+			},
+			{
+				Key:   []byte("sender"),
+				Value: []byte(sender.String()),
+			},
+		},
 	}
 }